LPD-22166 Move create license to parent license folder and show licenses tab for trials and paid orders
