+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/CustomerDashboard/pages/Apps/App/Licenses/CreateLicense/SelectSubscription.tsx
@@ -8,10 +8,10 @@ import classNames from 'classnames';
 import {useParams} from 'react-router-dom';
 import useSWR from 'swr';
 
-import RadioCardList from '../../../components/RadioCardList/RadioCardList';
-import useMarketplaceSpringBootOAuth2 from '../../../hooks/useMarketplaceSpringBootOAuth2';
-import i18n from '../../../i18n';
-import {formatDate} from '../../PublisherDashboard/PublisherDashboardPageUtil';
+import RadioCardList from '../../../../../../../components/RadioCardList/RadioCardList';
+import useMarketplaceSpringBootOAuth2 from '../../../../../../../hooks/useMarketplaceSpringBootOAuth2';
+import i18n from '../../../../../../../i18n';
+import {formatDate} from '../../../../../../PublisherDashboard/PublisherDashboardPageUtil';
 
 type SubscriptionSelectionProps = {
 	onSelectSubscription: (subscription: any) => void;
@@ -23,9 +23,7 @@ const SelectSubscription = ({
 	selectedSubscriptionValue,
 }: SubscriptionSelectionProps) => {
 	const params = useParams();
-
 	const orderId = Number(params.orderId);
-
 	const marketplaceSpringBootOAuth2 = useMarketplaceSpringBootOAuth2();
 
 	const {
