+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/CustomerDashboard/pages/Apps/App/Licenses/CreateLicense/Types/index.ts
@@ -17,7 +17,7 @@ export type CreateLicenseForm = {
 	subscription: any;
 };
 
-export type StepsInformationProps = {
+type StepsInformationProps = {
 	backStep: string;
 	nextStep: string;
 	stepTitle: string;
