+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/CustomerDashboard/pages/Apps/App/Licenses/Licenses.tsx
@@ -30,10 +30,10 @@ import {OrderType} from '../../../../../../enums/OrderType';
 import useGetProductByOrderId from '../../../../../../hooks/useGetProductByOrderId';
 import useMarketplaceSpringBootOAuth2 from '../../../../../../hooks/useMarketplaceSpringBootOAuth2';
 import {LicenseKey} from '../../../../../../services/oauth/MarketplaceSpringBootOAuth2';
-import LicenseDetailsModalHeader from './components/LicenseDetailsModalHeader';
-import LicenceKeyModalContent from './components/LicenseModalContent';
-import TableActions from './components/TableActions';
-import TitleSubtitleHeader from './components/TitleSubtitleHeader';
+import LicenseDetailsModalHeader from '../../../../components/LicenseDetailsModalHeader';
+import LicenceKeyModalContent from '../../../../components/LicenseModalContent';
+import TableActions from '../../../../components/TableActions';
+import TitleSubtitleHeader from '../../../../components/TitleSubtitleHeader';
 import useLicenseActions from './useLicensesActions';
 
 type OutletContext = ReturnType<typeof useGetProductByOrderId>;
